added tests for presence of ss:data
I just discovered that Excel will output empty cells if you copy and
paste from one worksheet to another, even if that cell is empty (even
though empty cells are not output when creating an Excel file from
scratch). Still, the empty cells won't have any Data, so I just needed
to add a check for the presence of that element.