Minor change to how dates are added
I'm going to add another style sheet to this directory that will add
textual date notes when those are missing for ASpace.  The problem is
that we don't want a date like <unitdate normal="1912"/> showing up as
"1912-1912" in ASpace after importing a file.  It shouldn't, but that's
what the EAD importer currently does.